boardid: Minor clean up and standardization
Merge the different coreboot table strapping ID structures into one
because they're really just all the same, and I want to add more. Make
the signature of the board_id() function return a uint32_t because
that's also what goes in the coreboot table. Add a printk to the generic
code handling strapping IDs in ramstage so that not every individual
mainboard implementation needs its own print. (In turn, remove one such
print from fsp1_1 code because it's in the way of my next patch.)

Reviewed-on: https://review.coreboot.org/22741
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/mainboard/google/foster/boardid.c
@@ -18,7 +18,7 @@
 #include <console/console.h>
 #include <gpio.h>
 
-uint8_t board_id(void)
+uint32_t board_id(void)
 {
 	return 0;
 }
