+++ src/mainboard/google/veyron/boardid.c
@@ -18,7 +18,7 @@
 #include <gpio.h>
 #include <stdlib.h>
 
-uint8_t board_id(void)
+uint32_t board_id(void)
 {
 	static int id = -1;
 	gpio_t pins[] = {[3] = GPIO(2, A, 7), [2] = GPIO(2, A, 2),
