+++ src/mainboard/google/zoombini/boardid.c
@@ -17,7 +17,7 @@
 #include <boardid.h>
 #include <stddef.h>
 
-uint8_t board_id(void)
+uint32_t board_id(void)
 {
 	MAYBE_STATIC int id = -1;
